Time: 43 ms (70.99%), Space: 61 MB (64.09%) - LeetHub

+++ 0493-reverse-pairs/0493-reverse-pairs.java
@@ -3,7 +3,7 @@ class Solution {
     public static void inversions(int[] a,int[] b){
         int i=0,j=0;
         while(i<a.length && j<b.length){
-            if((long)a[i]>(long)2*b[j]){
+            if(a[i]>(long)2*b[j]){
                 count += (a.length - i);
                 j++;
             }
